+++ index.js
@@ -5,4 +5,3 @@
 
 export { default as remarkMddDocumentStructure } from './plugins/remark-mdd-document-structure.js'
 export { default as remarkMddTextFormatting } from './plugins/remark-mdd-text-formatting.js'
-export { default as remarkMdxConditional } from './plugins/remark-mdx-conditional.js'